+++ app/controllers/locations_controller.rb
@@ -1,5 +1,4 @@
 class LocationsController < ApplicationController
-  load_and_authorize_resource
   
   def index
     render json: Location.all
@@ -7,6 +6,7 @@ class LocationsController < ApplicationController
 
   def create
     location = Location.new(location_params)
+    authorize location
     
     if location.save
       render json: location, status: :ok
@@ -17,6 +17,7 @@ class LocationsController < ApplicationController
 
   def update
     location = Location.find(params[:id])
+    authorize location
 
     if location.update(location_params)
       render json: location, status: :ok
@@ -27,6 +28,7 @@ class LocationsController < ApplicationController
 
   def destroy
     location = Location.find(params[:id])
+    authorize location
 
     if location.destroy
       render json: :ok
